requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
opencv-python>=4.7.0
scikit-image>=0.20.0
matplotlib>=3.6.0
pandas>=1.5.0
numpy>=1.24.0
av>=10.0.0